Update admin pages and backend also added list page

+++ backend/config/cloudinary.js
@@ -1,28 +1,32 @@
-import { v2 as cloudinary } from 'cloudinary';
-import fs from 'fs'
-
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
 const uploadOnCloudinary = async (filePath) => {
-    cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY , 
-        api_secret: process.env.CLOUDINARY_API_SECRET 
-    });
-    try {
-        if(!filePath){
-        return null
-    }
-    const uploadResult = await cloudinary.uploader.upload
-    (filePath)
-    fs.unlinkSync(filePath)
-    return uploadResult.secure_url
+  cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+  });
 
-    
-        
-    } catch (error) {
-        fs.unlinkSync(filePath)
-        console.log(error)
+  try {
+    if (!filePath) {
+      return null;
     }
-    
-}
-export default uploadOnCloudinary
+
+    const uploadResult = await cloudinary.uploader.upload(filePath);
+
+    fs.unlinkSync(filePath);
+
+    return uploadResult.secure_url;
+  } catch (error) {
+    console.log("Cloudinary Upload Error:", error);
+
+    if (filePath && fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
+
+    return null;
+  }
+};
+
+export default uploadOnCloudinary;
